Move asn1 defs to own file, move non-x509 stuff to pkcs7.c
