+++ crypto/CMakeLists.txt
@@ -467,6 +467,7 @@ add_library(
   pem/pem_x509.c
   pem/pem_xaux.c
   pkcs7/pkcs7.c
+  pkcs7/pkcs7_asn1.c
   pkcs7/pkcs7_x509.c
   pkcs8/pkcs8.c
   pkcs8/pkcs8_x509.c
